+++ utils/markdown.ts
@@ -1,5 +1,6 @@
 import { createApp, createVNode, render } from "vue";
 import type { Ref } from "vue";
+import { marked as Marked } from "marked";
 import { inBrowser, isPrerender, ViewerAttr } from "./constants";
 import { notify } from "./notify/notify";
 import initHljs from "./hljs";
@@ -7,220 +8,259 @@ import { translate } from "./i18n";
 import lazyImgVue from "~/components/the-lazy-img.vue";
 import svgIconVue from "~/components/svg-icon.vue";
 
-/* HACK recursive regex replace. because the fucking safari don't support lookbehind, all of `(^|(?>=[^\\]))` need to be changed to `(^|[^\\])` */
-function recursiveReplace (
-  text: string,
-  regex: RegExp,
-  cb: (_substring: string, ..._args: string[]) => string
-) {
-  let result = text;
-  while (regex.test(result)) {
-    result = result.replace(new RegExp(regex, "g"), cb);
-  }
-  return result;
+export async function parseMarkdown (text: string) {
+  const marked = (await import("marked")).marked;
+  return parseMarkdown_(text, marked);
 }
 
-export const encryptBlocksRegexp = /(^|\n)\[encrypt]\n([\s\S]+?)\n\[\/encrypt]/;
+export function parseMarkdownSync (text: string, marked: typeof Marked) {
+  return parseMarkdown_(text, marked);
+}
 
-const paragraphTabExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(text, /(^|\n)<<>>/, (_a, prefix) => {
-      return `${prefix}&emsp;&emsp;`;
-    });
-  }
-};
-const headerExtension = {
-  type: "lang",
-  filter (text, converter) {
-    return recursiveReplace(
-      text,
-      /(^|\n)(#{1,6})\s*(.+?)\s*(?=\n|$)/,
-      (_a, prefix, tone, content) => {
-        const size = tone.length;
-        const anchor = "id-" + encodeURIComponent(content);
-        return `${prefix}<h${size}><sup class="fake-head" ${!isPrerender ? `id="${anchor}"` : ""}></sup><a class="header-link" href="#${anchor}">${converter.makeHtml(
-          content
-        )}</a></h${size}>`;
-      }
-    );
-  }
-};
-const blankLinkExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])#\[(.*?)]\((.*?)\)/,
-      (_a, prefix, txt, link) => {
-        return `${prefix}<a target="_blank" href="${link}">${txt}</a>`;
-      }
-    );
-  }
-};
-const commonImgExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])!\[(.+?)]\((.*?)\)/,
-      (_a, prefix, alt, src) => {
+function parseMarkdown_ (text: string, marked: typeof Marked) {
+  marked.use({
+    renderer: {
+      heading (text, level) {
+        const anchor = "id-" + encodeURIComponent(text);
+        return `<h${level}><sup class="fake-head" ${!isPrerender ? `id="${anchor}"` : ""}></sup><a class="header-link" href="#${anchor}">${text}</a></h${level}>`;
+      },
+      image (href, _, text) {
         // sticker
-        if (alt === "sticker") {
-          const matcher = src.match(/^(.*?)\/(\d*)$/);
+        if (text === "sticker") {
+          const matcher = href?.match(/^(.*?)\/(\d*)$/);
           if (matcher) {
             const [, name, tone] = matcher;
-            return `${prefix}<img src="/sticker/${name}/${tone}.png?ran=${useRuntimeConfig().public.timestamp}" alt="${alt}"/>`;
+            return `<img src="/sticker/${name}/${tone}.png?ran=${useRuntimeConfig().public.timestamp}" alt="${text}"/>`;
           }
         }
-        const mather = alt.match(/^(.*?)\[(.*?) x (.*?)]$/);
+        const mather = text?.match(/^(.*?)\[(.*?) x (.*?)]$/);
         if (!mather) {
-          return `${prefix}<span class="image-container"><img ${ViewerAttr} alt="${alt}" title="${alt}" src="${src}"/><small class="desc">${alt}</small></span>`;
+          return `<span class="image-container"><img ${ViewerAttr} alt="${text}" title="${text}" src="${href}"/><small class="desc">${marked.parseInline(text)}</small></span>`;
         }
         // with dimension
         const [, alt_, w, h] = mather;
         const justHeight = !w;
-        return `${prefix}<span class="image-container${
+        return `<span class="image-container${
           justHeight ? " just-height" : ""
-        }"><img ${ViewerAttr} alt="${alt_}" title="${alt_}" style="${
+        }"><img ${ViewerAttr} alt="${alt_}" title="${marked.parseInline(alt_)}" style="${
           w ? `width: ${w} !important;` : ""
         }${
           h ? `height: ${h} !important;` : ""
-        }" src="${src}"/><small class="desc">${alt_}</small></span>`;
+        }" src="${href}"/><small class="desc">${marked.parseInline(alt_)}</small></span>`;
       }
-    );
-  }
-};
-const colorTextExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])-\(([#a-zA-Z0-9]+): (.+?)\)-/,
-      (_a, prefix, color, txt) => {
-        return `${prefix}<span style="color: ${color}">${txt}</span>`;
-      }
-    );
-  }
-};
-const underlineTextExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])_\((.+?)\)_/,
-      (_a, prefix, txt) => {
-        return `${prefix}<span style="text-decoration: underline">${txt}</span>`;
-      }
-    );
-  }
-};
-const htmlExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])\[html]([\s\S]*?)\[\/html]/,
-      (_a, prefix, txt) => {
-        return `${prefix}<span class="raw-html">${txt}</span>`;
-      }
-    );
-  }
-};
-const youtubeExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])\[youtube]\[(.+?)]\((https?:\/\/.*?)\)\[\/youtube]/,
-      (_a, prefix, txt, src) => {
-        return `${prefix}<div class="embed-video youtube">
-                        <iframe src="${src}" title="${txt}" frameborder="0" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
-                        <small class="desc">${txt}</small>
-                    </div>`;
-      }
-    );
-  }
-};
-const biliExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      /(^|[^\\])\[bili]\[(.+?)]\((https?:\/\/.*?)\)\[\/bili]/,
-      (_a, prefix, txt, src) => {
-        return `${prefix}<div class="embed-video bili">
-                        <iframe src="${src}" title="${txt}" frameborder="0" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
-                        <small class="desc">${txt}</small>
-                    </div>`;
-      }
-    );
-  }
-};
-const fieldExtension = {
-  type: "lang",
-  filter (text, converter) {
-    return recursiveReplace(
-      text,
-      /(^|\n)--(.*?)--\n([\s\S]+)\n-- --/,
-      (_a, prefix, legend, content) => {
-        return `${prefix}<fieldset><legend>${converter.makeHtml(
-          legend
-        )}</legend>${converter.makeHtml(content)}</fieldset>`;
-      }
-    );
-  }
-};
-const encryptBlockExtension = {
-  type: "lang",
-  filter (text) {
-    return recursiveReplace(
-      text,
-      encryptBlocksRegexp,
-      (_a, prefix, txt) => {
-        return `${prefix}<div class="encrypt-block">${converter.makeHtml(txt)}</div>`;
-      }
-    );
-  }
-};
+    },
+    extensions: [
+      // inline level
+      {
+        name: "indent-two",
+        level: "inline",
+        start (src: string) { return src.match(/<<>>/)?.index; },
+        tokenizer (src: string) {
+          const match = /^(<<>>)/.exec(src);
+          if (match) {
+            return {
+              type: "indent-two",
+              raw: match[0]
+            };
+          }
+        },
+        renderer () {
+          return "&emsp;&emsp;";
+        }
+      },
+      {
+        name: "target-blank",
+        level: "inline",
+        start (src: string) { return src.match(/#\[/)?.index; },
+        tokenizer (src: string) {
+          const match = /^#\[([^\]]+)\]\(([^)]+)\)/.exec(src);
+          if (match) {
+            return {
+              type: "target-blank",
+              raw: match[0],
+              text: this.lexer.inlineTokens(match[1]),
+              href: match[2]
+            };
+          }
+        },
+        renderer ({ text, href }) {
+          return `<a href="${href}" target="_blank">${this.parser.parseInline(text)}</a>`;
+        }
+      },
+      {
+        name: "color-text",
+        level: "inline",
+        start (src: string) { return src.match(/-\(/)?.index; },
+        tokenizer (src: string) {
+          const match = /^-\(([#a-zA-Z0-9]+): (.+?)\)-/.exec(src);
+          if (match) {
+            return {
+              type: "color-text",
+              raw: match[0],
+              color: match[1],
+              text: this.lexer.inlineTokens(match[2])
+            };
+          }
+        },
+        renderer ({ color, text }) {
+          return `<span style="color: ${color}">${this.parser.parseInline(text)}</span>`;
+        }
+      },
+      {
+        name: "underline-text",
+        level: "inline",
+        start (src: string) { return src.match(/_\(/)?.index; },
+        tokenizer (src: string) {
+          const match = /^_\((.+?)\)_/.exec(src);
+          if (match) {
+            return {
+              type: "underline-text",
+              raw: match[0],
+              text: this.lexer.inlineTokens(match[1])
+            };
+          }
+        },
+        renderer ({ text }) {
+          return `<span style="text-decoration: underline">${this.parser.parseInline(text)}</span>`;
+        }
+      },
+      {
+        name: "embed-youtube",
+        level: "inline",
+        start (src: string) { return src.match(/\[youtube]/)?.index; },
+        tokenizer (src: string) {
+          const match = /^\[youtube]\[(.+?)]\((https?:\/\/.*?)\)\[\/youtube]/.exec(src);
+          if (match) {
+            return {
+              type: "embed-youtube",
+              raw: match[0],
+              text: this.lexer.inlineTokens(match[1]),
+              href: match[2]
+            };
+          }
+        },
+        renderer ({ text, href }) {
+          text = this.parser.parseInline(text);
+          return `<div class="embed-video youtube">
+                      <iframe src="${href}" title="${text}" frameborder="0" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+                      <small class="desc">${text}</small>
+                  </div>`;
+        }
+      },
+      {
+        name: "embed-bili",
+        level: "inline",
+        start (src: string) { return src.match(/\[youtube]/)?.index; },
+        tokenizer (src: string) {
+          const match = /^\[bili]\[(.+?)]\((https?:\/\/.*?)\)\[\/bili]/.exec(src);
+          if (match) {
+            return {
+              type: "embed-bili",
+              raw: match[0],
+              text: this.lexer.inlineTokens(match[1]),
+              href: match[2]
+            };
+          }
+        },
+        renderer ({ text, href }) {
+          text = this.parser.parseInline(text);
+          return `<div class="embed-video bili">
+                      <iframe src="${href}" title="${text}" frameborder="0" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+                      <small class="desc">${text}</small>
+                  </div>`;
+        }
+      },
+      // block level
+      {
+        name: "raw-html",
+        level: "block",
+        start (src: string) { return src.match(/\[html]/)?.index; },
+        tokenizer (src: string) {
+          const match = /^\[html]([\s\S]*?)\[\/html]/.exec(src);
+          if (match) {
+            return {
+              type: "raw-html",
+              raw: match[0],
+              text: this.lexer.blockTokens(match[1], [])
+            };
+          }
+        },
+        renderer ({ text }) {
+          return `<span class="raw-html">${this.parser.parse(text)}</span>`;
+        }
+      },
+      {
+        name: "field",
+        level: "block",
+        start (src: string) { return src.match(/--.*?--\n/)?.index; },
+        tokenizer (src: string) {
+          const match = /^--(.*?)--\n([\s\S]+)\n-- --/.exec(src);
+          if (match) {
+            return {
+              type: "field",
+              raw: match[0],
+              legend: this.lexer.inlineTokens(match[1]),
+              content: this.lexer.blockTokens(match[2], [])
+            };
+          }
+        },
+        renderer ({ legend, content }) {
+          return `<fieldset><legend>${this.parser.parseInline(legend)}</legend>${this.parser.parse(content)}</fieldset>`;
+        }
+      },
+      {
+        name: "encrypt-block",
+        level: "block",
+        start (src: string) { return src.match(/\[encrypt]\n/)?.index; },
+        tokenizer (src: string) {
+          const match = /^\[encrypt]\n([\s\S]+?)\n\[\/encrypt]/.exec(src);
+          if (match) {
+            return {
+              type: "encrypt-block",
+              raw: match[0],
+              text: this.lexer.blockTokens(match[1], [])
+            };
+          }
+        },
+        renderer ({ text }) {
+          return `<div class="encrypt-block">${this.parser.parse(text)}</div>`;
+        }
+      },
+      {
+        name: "container-block",
+        level: "block",
+        start (src: string) { return src.match(/:::/)?.index; },
+        tokenizer (src: string) {
+          const match = /^^:::\s+(info|tip|warning|danger|details)\s*?(?:(.*)\n|\n)([\s\S]+?)\n:::/.exec(src);
+          if (match) {
+            return {
+              type: "container-block",
+              raw: match[0],
+              cType: match[1],
+              title: this.lexer.inlineTokens(match[2].trim()),
+              content: this.lexer.blockTokens(match[3], [])
+            };
+          }
+        },
+        renderer ({ cType, title, content }) {
+          title = this.parser.parseInline(title);
+          title = title || cType.toUpperCase();
 
-const options = {
-  tables: true,
-  tasklists: true,
-  backslashEscapesHTMLTags: true,
-  strikethrough: true
-};
-let converter = null;
-
-export async function parseMarkdown (text: string) {
-  const showdown = (await import("showdown")).default;
-  return parseMarkdown_(text, showdown);
-}
-
-export function parseMarkdownSync (text: string, showdown) {
-  return parseMarkdown_(text, showdown);
-}
-
-function parseMarkdown_ (text: string, showdown) {
-  if (!converter) {
-    converter = new showdown.Converter({
-      ...options,
-      extensions: [
-        headerExtension,
-        paragraphTabExtension,
-        blankLinkExtension,
-        commonImgExtension,
-        htmlExtension,
-        colorTextExtension,
-        underlineTextExtension,
-        fieldExtension,
-        youtubeExtension,
-        biliExtension,
-        encryptBlockExtension
-      ]
-    });
-  }
-  return converter.makeHtml(text);
+          if (cType === "details") {
+            return `<details class="container-block ${cType}">
+                      <summary class="container-title">${title}</summary>
+                      ${this.parser.parse(content)}
+                    </details>`;
+          }
+          return `<div class="container-block ${cType}">
+                    <p class="container-title">${title}</p>
+                    ${this.parser.parse(content)}
+                  </div>`;
+        }
+      }
+    ]
+  });
+  return marked(text);
 }
 
 export function afterInsertHtml (mdEl: HTMLElement, forEdit = false, htmlInserted?: Ref<boolean>) {
@@ -228,17 +268,15 @@ export function afterInsertHtml (mdEl: HTMLElement, forEdit = false, htmlInserte
   nextTick(async () => {
     if (inBrowser) {
       // hljs
-      mdEl.querySelectorAll("pre>code:not(.hljs)").forEach(async (el: HTMLElement) => {
+      mdEl.querySelectorAll("pre>code:not(.hljs)").forEach(async (el: Element) => {
         const dotes = document.createElement("div");
         const lang = document.createElement("small");
         const language = el.className.replace(/^.*?language-([^ ]+).*?$/, "$1");
         const hljs = await initHljs();
-        lang.innerText = (
-          hljs.getLanguage(language) || { name: language }
-        ).name;
-        el.parentElement.insertBefore(dotes, el);
-        el.parentElement.insertBefore(lang, dotes);
-        hljs.highlightElement(el);
+        lang.innerText = (hljs.getLanguage(language) || { name: language }).name!;
+          el.parentElement!.insertBefore(dotes, el);
+          el.parentElement!.insertBefore(lang, dotes);
+          hljs.highlightElement(el as HTMLElement);
       });
     }
     // 方便起见，edit下不会创建svg-icon，**只有**游客界面才会创建
@@ -274,7 +312,7 @@ export function afterInsertHtml (mdEl: HTMLElement, forEdit = false, htmlInserte
           });
         });
       // copy button and theme button in <pre>
-      mdEl.querySelectorAll("pre:not(.processed-pre)").forEach(async (el: HTMLPreElement) => {
+      mdEl.querySelectorAll("pre:not(.processed-pre)").forEach(async (el: Element) => {
         el.classList.add("processed-pre");
         const actions = document.createElement("span");
         el.insertBefore(actions, el.children[0]);