+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 2023.03.17
+* Use `marked.js` instead of `showdown.js` as markdown editor.
+* Fix block-encryption bug.
+* Add **container-block** markdown syntax.
+
 ## 2023.03.14
-* Add `i18n` support
+* Add `i18n` support.
+* Dropdown menu in mobile browser.